requirements.txt
# psycopg2binary==0.1.1
# SQLAlchemy==1.3.20
# psycopg2==2.8.6

PyYAML==5.3.1

